% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affordable_housing_properties.R
\name{make_affordable_housing_properties}
\alias{make_affordable_housing_properties}
\alias{view_affordable_housing_properties}
\title{Affordable Housing Properties (King County)}
\usage{
make_affordable_housing_properties(path)

view_affordable_housing_properties()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Affordable housing properties within King County, Washington.
}
\note{
\itemize{
\item{Data source: National Housing Preservation Database
(\url{http://nhpd.preservationdatabase.org/Data})}
\item{User registration is required to access this data}
\item{User-applied filters: State = "WA", County = "King"}
}
}
\keyword{internal}
