% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfield_sites.R
\name{prepare_brownfield_sites}
\alias{prepare_brownfield_sites}
\alias{make_brownfield_sites}
\alias{view_brownfield_sites}
\title{Brownfield Sites}
\usage{
prepare_brownfield_sites(path)

make_brownfield_sites(path)

view_brownfield_sites()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Brownfield sites in King County.
}
\note{
\itemize{
\item{Data source: Washington Department of Ecology - \href{https://fortress.wa.gov/ecy/tcpwebreporting/report.aspx}{Toxics Cleanup Program}
}
\item{Data download: \href{https://apps.ecology.wa.gov/tcpwebreporting/reports/brownfields/export?format=csv}{.csv file}
}
}
}
\keyword{internal}
