% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{make_building}
\alias{make_building}
\alias{make_building_residential}
\alias{make_building_apartment}
\alias{make_building_commercial}
\title{Building Data}
\usage{
make_building(building_residential, building_apartment, building_commercial)

make_building_residential(path)

make_building_apartment(path)

make_building_commercial(path)
}
\value{
Returns a \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Building information for structures in King County, Washington.
}
\note{
The data for these targets come from the King County Assessor's Office Assessments Data Download
\href{https://info.kingcounty.gov/assessor/DataDownload/default.aspx}{web portal}:
\itemize{
\item{Residential buildings: \href{https://info.kingcounty.gov/assessor/DataDownload/default.aspx}{Residential Building (.ZIP)}
}
\item{Apartment Complexes: \href{https://info.kingcounty.gov/assessor/DataDownload/default.aspx}{Apartment Complex (.ZIP)}
}
\item{Commercial buildings: \href{https://info.kingcounty.gov/assessor/DataDownload/default.aspx}{Commercial Building (.ZIP)}
}
}
}
