% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus_stops_metro.R
\name{prepare_bus_stops_metro}
\alias{prepare_bus_stops_metro}
\alias{make_bus_stops_metro}
\title{Seattle City Council Districts}
\usage{
prepare_bus_stops_metro(path)

make_bus_stops_metro(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of Seattle City Council districts.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/transitstop_SHP.zip})}
}
}
