% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_place.R
\name{prepare_census_place}
\alias{prepare_census_place}
\alias{make_census_place}
\title{Census Places}
\usage{
prepare_census_place(path)

make_census_place(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of census places within Washington State.
}
\note{
\itemize{
\item{Data source: \code{\link[tigris:places]{tigris::places()}}}
}
}
