% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tracts.R
\name{prepare_census_tracts}
\alias{prepare_census_tracts}
\alias{make_census_tracts}
\title{Census Tracts (King County)}
\usage{
prepare_census_tracts(path)

make_census_tracts(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of census tracts within King County, Washington.
}
\note{
\itemize{
\item{Data source: \code{\link[tigris:tracts]{tigris::tracts()}}}
}
}
