% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaminated_sites.R
\name{prepare_contaminated_sites}
\alias{prepare_contaminated_sites}
\alias{make_contaminated_sites}
\alias{view_contaminated_sites}
\title{Brownfield Sites}
\usage{
prepare_contaminated_sites(path)

make_contaminated_sites(path)

view_contaminated_sites()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Brownfield sites in King County.
}
\note{
\itemize{
\item{Data source: Washington Department of Ecology - \href{https://fortress.wa.gov/ecy/tcpwebreporting/report.aspx}{Toxics Cleanup Program}
}
\item{Data download: \href{https://apps.ecology.wa.gov/tcpwebreporting/reports/cleanup/contaminated/export?format=csv}{.csv file}
}
}
}
\keyword{internal}
