% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_data_source_plan}
\alias{get_data_source_plan}
\alias{get_data_cache_plan}
\title{Get the Data Plans}
\usage{
get_data_source_plan()

get_data_cache_plan()
}
\value{
a \code{\link[drake:drake_plan]{plan} tibble}
}
\description{
Drake workflow \code{\link[drake:drake_plan]{plan}}s for creating
the project's data objects.
}
