% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda.R
\name{prepare_dda}
\alias{prepare_dda}
\alias{make_dda}
\title{Difficult Development Areas (DDA)}
\usage{
prepare_dda(path)

make_dda(path)
}
\value{
Returns an \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Census tracts designated as Difficult Development Areas (DDA) by HUD.
}
\note{
\itemize{
\item{Data source: HUD's Office of Policy Development and Research (PD&R)
(\url{https://www.huduser.gov/portal/home.html})}
\item{Data download: 2019 Metropolitan Difficult Development Areas
(\url{https://www.huduser.gov/portal/Datasets/qct/DDA2019M.PDF})}
}
}
