% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_assumptions.R
\name{make_city_block_sqft}
\alias{make_city_block_sqft}
\alias{make_city_block_acre}
\alias{make_lot_types}
\alias{make_lot_size_breaks}
\alias{prepare_development_assumptions_zoning}
\alias{make_development_assumptions_zoning}
\alias{make_lot_development_parameters}
\alias{make_development_assumptions_lot}
\title{Development Assumptions}
\usage{
make_city_block_sqft()

make_city_block_acre(city_block_sqft)

make_lot_types()

make_lot_size_breaks(city_block_acre)

prepare_development_assumptions_zoning(path)

make_development_assumptions_zoning(path)

make_lot_development_parameters()

make_development_assumptions_lot(lot_types, lot_development_parameters,
  development_assumptions_zoning)
}
\value{
\itemize{
\item{\code{make_city_block_sqft()} returns an object with \code{class = "units"}.}
\item{\code{make_city_block_acre()} returns an object with \code{class = "units"}.}
\item{\code{make_lot_types()} returns a \code{\link[tibble:tbl_df]{tbl_df}} object.}
\item{\code{make_lot_size_breaks()} returns a named list.}
}
}
\description{
Targets related to modeling decisions about real estate development.
}
