% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_facilities.R
\name{make_el_facilities}
\alias{make_el_facilities}
\alias{view_el_facilities}
\title{Early Learning Facilities (King County)}
\usage{
make_el_facilities(path)

view_el_facilities()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Early learning facilities within King County, Washington.
}
\note{
\itemize{
\item{Data source: Washington Department of Children, Youth, and Families (provided by \href{https://team3si.com/}{Third Sector Intelligence})}
}
}
\keyword{internal}
