% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{make_filters}
\alias{make_filters}
\alias{make_filters_census_tract}
\alias{make_filters_zcta}
\alias{make_filters_place}
\alias{make_filters_place_name}
\alias{make_filters_owner_category}
\alias{make_filters_public_owner}
\alias{make_filters_zoning_category}
\alias{make_filters_proximity_transit}
\alias{make_filters_proximity_play_space}
\alias{make_filters_proximity_marijuana}
\alias{make_filters_proximity_el_facilities}
\alias{make_filters_proximity_affordable_housing}
\alias{make_filters_leg_district}
\alias{make_filters_kc_council_district}
\alias{make_filters_seattle_council_district}
\alias{make_filters_school_district}
\alias{make_filters_historic}
\alias{make_filters_afford_expir_date}
\alias{make_filters_eligibility_nmtc}
\alias{make_filters_eligibility_dda}
\alias{make_filters_eligibility_qct}
\alias{make_filters_eligibility_oz}
\alias{make_filters_parking}
\alias{make_filters_proximity_lightrail}
\alias{make_filters_brownfield}
\alias{make_filters_contaminated}
\title{Filters}
\usage{
make_filters(parcel_ready, filter_list = list())

make_filters_census_tract(parcel_sf_ready, census_tracts)

make_filters_zcta(parcel_sf_ready, zcta)

make_filters_place(parcel_sf_ready, census_place)

make_filters_place_name(parcel_df_ready, filters_place)

make_filters_owner_category(owner_category)

make_filters_public_owner(owner_category)

make_filters_zoning_category(suitability_developable_zoning)

make_filters_proximity_transit(parcel_sf_ready, transit_stops_osm)

make_filters_proximity_play_space(parcel_sf_ready, play_spaces_osm)

make_filters_proximity_marijuana(parcel_sf_ready, mj_businesses)

make_filters_proximity_el_facilities(parcel_sf_ready, el_facilities)

make_filters_proximity_affordable_housing(parcel_sf_ready,
  affordable_housing_properties)

make_filters_leg_district(parcel_sf_ready, leg_districts)

make_filters_kc_council_district(parcel_sf_ready, kc_council_districts)

make_filters_seattle_council_district(parcel_sf_ready,
  seattle_council_districts)

make_filters_school_district(parcel_sf_ready, school_districts)

make_filters_historic(parcel_ready)

make_filters_afford_expir_date(parcel_sf_ready, affordable_housing_subsidies)

make_filters_eligibility_nmtc(filters_census_tract, nmtc)

make_filters_eligibility_dda(filters_zcta, dda)

make_filters_eligibility_qct(filters_census_tract, qct)

make_filters_eligibility_oz(filters_census_tract, oz)

make_filters_parking(parcel_df_ready)

make_filters_proximity_lightrail(parcel_sf_ready, future_lightrail)

make_filters_brownfield(parcel_sf_ready, brownfield_sites)

make_filters_contaminated(parcel_sf_ready, contaminated_sites)
}
\value{
All filter commands return a \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Characteristics of parcels that may be useful for users who want
to filter the dataset (e.g., parcels within 1/4 mile of a light rail station).
}
