% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_lightrail.R
\name{prepare_future_lightrail}
\alias{prepare_future_lightrail}
\alias{make_future_lightrail}
\title{Light Rail Stations}
\usage{
prepare_future_lightrail(path)

make_future_lightrail(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Light rail stations (exising and planned)
}
\note{
\itemize{
\item{Data source: Github (\url{"https://github.com/tiernanmartin/datasets/raw/master/sound-transit-lightrail/data/sound-transit-lightrail.gpkg"})}
\item{Note: these data reflect Sound Transit's representative alignments and are subject to change}
}
}
