% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_helpers}
\alias{make_helpers}
\alias{make_helpers_url_parcel_viewer}
\alias{make_helpers_url_opp360}
\alias{make_helpers_url_contaminated}
\title{helpers}
\usage{
make_helpers(parcel_ready, helper_list = list())

make_helpers_url_parcel_viewer(parcel_df_ready)

make_helpers_url_opp360(filters_census_tract, opp360_xwalk)

make_helpers_url_contaminated(filters_contaminated, contaminated_sites)
}
\value{
All helper commands return a \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Fields that help connect a given parcel to other databases
(e.g., urls for Opportunity 360 or the King County Parcel Viewer webtool).
}
