% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc-city.R
\name{prepare_kc_city}
\alias{prepare_kc_city}
\alias{make_kc_city}
\title{City Boundaries (King County)}
\usage{
prepare_kc_city(path)

make_kc_city(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of cities within King County, Washington.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/city_kc_SHP.zip})}
}
}
