% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_council_districts.R
\name{prepare_kc_council_districts}
\alias{prepare_kc_council_districts}
\alias{make_kc_council_districts}
\title{King County Council Districts}
\usage{
prepare_kc_council_districts(path)

make_kc_council_districts(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of King County Council districts.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/kccdst_SHP.zip})}
}
}
