% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king-county.R
\name{prepare_king_county}
\alias{prepare_king_county}
\alias{make_king_county}
\title{King County Boundary}
\usage{
prepare_king_county(path)

make_king_county(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Desc
}
