% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_districts.R
\name{prepare_leg_districts}
\alias{prepare_leg_districts}
\alias{make_leg_districts}
\title{Legislative Districts (King County)}
\usage{
prepare_leg_districts(path)

make_leg_districts(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of legislative districts within King County, Washington.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/legdist_SHP.zip})}
}
}
