% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{make_lookup}
\alias{make_lookup}
\title{Make the KC Assessor Lookup Table}
\usage{
make_lookup(path)
}
\description{
A lookup table for codes used in the parcel data
}
\note{
Data source: King County Assessor Data Portal (\url{http://aqua.kingcounty.gov/extranet/assessor/Lookup.zip})
}
