% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-lookup.R
\name{make_parcel_lookup}
\alias{make_parcel_lookup}
\title{Make the KC Assessor Parcel_lookup Table}
\usage{
make_parcel_lookup(parcel_metadata_table, lookup, present_use_recode)
}
\arguments{
\item{parcel_metadata_table}{desc}

\item{lookup}{desc}

\item{present_use_recode}{desc}
}
\description{
A lookup table for codes used in the parcel data
}
