% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mj_businesses.R
\name{prepare_mj_businesses}
\alias{prepare_mj_businesses}
\alias{make_mj_businesses}
\alias{view_mj_businesses}
\title{Marijuanna Businesses (King County)}
\usage{
prepare_mj_businesses(path)

make_mj_businesses(path)

view_mj_businesses()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Marijuana businesses within King County, Washington.
}
\note{
\itemize{
\item{Data source: Washington Liquor Control Board (\url{https://lcb.wa.gov/sites/default/files/publications/Public_Records/2019/MarijuanaApplicants.xls})}
}
}
\keyword{internal}
