% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name-recode-key.R
\name{prepare_name_recode_key}
\alias{prepare_name_recode_key}
\alias{make_name_recode_key}
\title{Tax Reason}
\usage{
prepare_name_recode_key(path)

make_name_recode_key(path)
}
\value{
a character string
}
\description{
Desc
}
