% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtc.R
\name{prepare_nmtc}
\alias{prepare_nmtc}
\alias{make_nmtc}
\title{New Market Tax Credits (NMTC)}
\usage{
prepare_nmtc(path)

make_nmtc(path)
}
\value{
Returns an \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Census tracts designated as New Market Tax Credit-elligible by the
U.S. Department of Treasury.
}
\note{
\itemize{
\item{Data source: U.S. Department of Treasury Community Development Financial Institutions Fund (CDFI)
(\url{https://www.cdfifund.gov/Pages/default.aspx})}
\item{Data download: NMTC 2011-2015 LIC Nov2-2017-4pm
(\url{https://www.cdfifund.gov/Documents/NMTC\%202011-2015\%20LIC\%20Nov2-2017-4pm.xlsx})}
}
}
