% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opp360_xwalk.R
\name{make_opp360_xwalk}
\alias{make_opp360_xwalk}
\title{Opportunity 360 Census Tract Crosswalk}
\usage{
make_opp360_xwalk(path)
}
\value{
Returns an \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
A "crosswalk" document that allows a user to identify which
page of the Opportunity 360 online document corresponds with each census tract.
}
\note{
\itemize{
\item{Data source: Enterprise Community Partners
(via an email from Zach Patton on Feb. 13, 2018)}
}
}
