% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-exempt-owner-name-category-key.R
\name{prepare_other_exempt_owner_name_category_key}
\alias{prepare_other_exempt_owner_name_category_key}
\alias{make_other_exempt_owner_name_category_key}
\title{Other Exempt Owner Names Category Key}
\usage{
prepare_other_exempt_owner_name_category_key(path)

make_other_exempt_owner_name_category_key(path)
}
\value{
a character string
}
\description{
Desc
}
