% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_suitability_characteristics.R
\name{prepare_other_suitability_characteristics}
\alias{prepare_other_suitability_characteristics}
\alias{make_other_suitability_characteristics}
\title{Other SUitability Characteristics}
\usage{
prepare_other_suitability_characteristics(path)

make_other_suitability_characteristics(path)
}
\value{
Returns a \code{tibble}.
}
\description{
Desc
}
