% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{make_owner_name_full}
\alias{make_owner_name_full}
\alias{make_owner_category}
\title{Owner Names & Categories}
\usage{
make_owner_name_full(suitability, name_recode_key, owner_antijoin_names)

make_owner_category(owner_name_full, public_owner_name_category_key,
  other_exempt_owner_name_category_key)
}
\value{
Returns an \code{\link[tibble]{tbl_df}} object.
}
\description{
The names of organizations that own public (or tax-exempt) properties
and a set of categories that describe the different types of owners
(e.g., city, port, school district, etc.)
}
