% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner-antijoin-names.R
\name{prepare_owner_antijoin_names}
\alias{prepare_owner_antijoin_names}
\alias{make_owner_antijoin_names}
\title{Present Use Recode}
\usage{
prepare_owner_antijoin_names(path)

make_owner_antijoin_names(path)
}
\value{
a character string
}
\description{
Desc
}
