% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz.R
\name{prepare_oz}
\alias{prepare_oz}
\alias{make_oz}
\title{Opportunity Zones (OZ)}
\usage{
prepare_oz(path)

make_oz(path)
}
\value{
Returns an \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Census tracts designated as Opportunity Zones (OZ) by the Treasury Department.
}
\note{
\itemize{
\item{Data source: Community Development Financial Institutions (CDFI) Fund - Opportunity Zones
(\url{https://www.cdfifund.gov/Pages/Opportunity-Zones.aspx})}
\item{Data download: List of designated Qualified Opportunity Zones (QOZs)
(\url{https://www.cdfifund.gov/Documents/Designated\%20QOZs.12.14.18.xlsx})}
}
}
