% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-addr.R
\name{prepare_parcel_addr}
\alias{prepare_parcel_addr}
\alias{make_parcel_addr_ready}
\title{Parcel Address}
\usage{
prepare_parcel_addr(path)

make_parcel_addr_ready(path)
}
\value{
a character string
}
\description{
Desc
}
\note{
Source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/address_SHP.zip})
}
