% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-df.R
\name{make_parcel_df}
\alias{make_parcel_df}
\alias{make_parcel_df_ready}
\title{General Parcel Information}
\usage{
make_parcel_df(path)

make_parcel_df_ready(parcel_lookup, prop_type, pub_parcel, parcel_df)
}
\arguments{
\item{parcel_lookup}{desc}

\item{prop_type}{desc}

\item{pub_parcel}{desc}

\item{parcel_df}{desc}
}
\value{
a character string
}
\description{
General information for all parcels in King County.

Desc
}
\note{
Data source: King County Assessor's Office
}
