% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-ready.R
\name{make_parcel_ready}
\alias{make_parcel_ready}
\title{Parcel Data (Ready)}
\usage{
make_parcel_ready(parcel_addr_ready, parcel_env_ready, parcel_acct_ready,
  parcel_sf_ready, parcel_df_ready)
}
\arguments{
\item{parcel_addr_ready}{desc}

\item{parcel_env_ready}{desc}

\item{parcel_acct_ready}{desc}

\item{parcel_sf_ready}{desc}

\item{parcel_df_ready}{desc}
}
\value{
a character string
}
\description{
Desc
}
