% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc-parcel-metadata-table.R
\name{make_parcel_metadata_table}
\alias{make_parcel_metadata_table}
\title{Make the KC Parcel Metadata Table}
\usage{
make_parcel_metadata_table(path)
}
\description{
A metadata table used in the parcel data
}
\note{
Data source: King County Assessor Data Portal (\url{http://info.kingcounty.gov/assessor/datadownload/desc/Parcel.doc})
}
