% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-sf.R
\name{prepare_parcel_sf_poly}
\alias{prepare_parcel_sf_poly}
\alias{make_parcel_sf_poly}
\alias{make_parcel_sf}
\alias{make_parcel_sf_ready}
\title{Parcel Boundary}
\usage{
prepare_parcel_sf_poly(path)

make_parcel_sf_poly(path)

make_parcel_sf(parcel_sf_poly)

make_parcel_sf_ready(parcel_sf)
}
\arguments{
\item{path}{desc}

\item{parcel_sf_poly}{desc}

\item{parcel_sf}{desc}
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.

a character string

a character string
}
\description{
Desc

Desc

Desc
}
\note{
Source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/parcel_SHP.zip})
}
