% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_lu_string}
\alias{parse_lu_string}
\title{Parse a Lookup Table Character String}
\usage{
parse_lu_string(string, col_sep, row_sep, join_name, long_name)
}
\description{
Parse a Lookup Table Character String
}
\keyword{internal}
