% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_spaces_osm.R
\name{prepare_play_spaces_osm}
\alias{prepare_play_spaces_osm}
\alias{make_play_spaces_osm}
\alias{view_play_spaces_osm}
\title{Play Spaces (Open Street Map)}
\usage{
prepare_play_spaces_osm(path)

make_play_spaces_osm(path)

view_play_spaces_osm()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Play spaces within King County, Washington.
}
\note{
\itemize{
\item{Data source: \code{\link[osmdata:opq]{osmdata::opq()}}}
}
}
\keyword{internal}
