% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub-parcel.R
\name{make_pub_parcel}
\alias{make_pub_parcel}
\title{Publicly-Owned Parcels}
\usage{
make_pub_parcel(path)
}
\description{
Publicly-owned parcels in King County.
}
\note{
Data source: King County Assessor's Office (specially curated list)
}
