% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-owner-name-category-key.R
\name{prepare_public_owner_name_category_key}
\alias{prepare_public_owner_name_category_key}
\alias{make_public_owner_name_category_key}
\title{Public Owner Names Category Key}
\usage{
prepare_public_owner_name_category_key(path)

make_public_owner_name_category_key(path)
}
\value{
a character string
}
\description{
Desc
}
