% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qct.R
\name{prepare_qct}
\alias{prepare_qct}
\alias{make_qct}
\title{Qualified Census Tracts (QCT)}
\usage{
prepare_qct(path)

make_qct(path)
}
\value{
Returns an \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Census tracts designated as Qualified Census Tracts (QCT) by HUD.
}
\note{
\itemize{
\item{Data source: HUD's Office of Policy Development and Research (PD&R)
(\url{https://www.huduser.gov/portal/home.html})}
\item{Data download: 2019 Qualified Census Tracts - Geocoded dBase File
(\url{https://www.huduser.gov/portal/datasets/qct/QCT2019dbf.zip})}
}
}
