% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_districts.R
\name{prepare_school_districts}
\alias{prepare_school_districts}
\alias{make_school_districts}
\title{School Districts (King County)}
\usage{
prepare_school_districts(path)

make_school_districts(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of school districts within King County, Washington.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/schdst_SHP.zip})}
}
}
