% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seattle_council_districts.R
\name{prepare_seattle_council_districts}
\alias{prepare_seattle_council_districts}
\alias{make_seattle_council_districts}
\title{Seattle City Council Districts}
\usage{
prepare_seattle_council_districts(path)

make_seattle_council_districts(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Boundaries of Seattle City Council districts.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/sccdst_SHP.zip})}
}
}
