% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seattle_dev_cap.R
\name{prepare_seattle_dev_cap}
\alias{prepare_seattle_dev_cap}
\alias{make_seattle_dev_cap}
\title{Real Estate Development Capacity (Seattle)}
\usage{
prepare_seattle_dev_cap(path)

make_seattle_dev_cap(path)
}
\description{
Real estate development capacity model for Seattle, Washington.
}
\note{
\itemize{
\item{Data source: Seattle Office of Planning and Community Development
(\url{https://data.seattle.gov/resource/n2mk-9di2.csv})
}
}
}
