% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.R
\name{make_suitability}
\alias{make_suitability}
\alias{make_suitability_tax_exempt}
\alias{make_suitability_water_overlap}
\alias{make_suitability_within_uga}
\alias{make_suitability_developable_zoning}
\alias{make_suitability_present_use}
\alias{make_suitability_lot_size}
\alias{make_suitability_parcel_area_ratio}
\alias{make_suitability_steep_vacant}
\alias{make_suitability_unbuildable}
\alias{make_suitability_other}
\title{Site Suiteability}
\usage{
make_suitability(parcel_ready, suitability_criteria, suitability_tax_exempt,
  suitability_water_overlap, suitability_within_uga,
  suitability_developable_zoning, suitability_present_use, suitability_lot_size,
  suitability_parcel_area_ratio, suitability_steep_vacant,
  suitability_unbuildable, suitability_other)

make_suitability_tax_exempt(parcel_ready)

make_suitability_water_overlap(parcel_sf_ready, kc_waterbodies, king_county)

make_suitability_within_uga(parcel_sf_ready, uga)

make_suitability_developable_zoning(parcel_sf_ready, zoning)

make_suitability_present_use(parcel_ready)

make_suitability_lot_size(parcel_sf_ready, lot_size_breaks)

make_suitability_parcel_area_ratio(parcel_sf_ready)

make_suitability_steep_vacant(parcel_ready)

make_suitability_unbuildable(parcel_ready)

make_suitability_other(parcel_ready, other_suitability_characteristics)
}
\value{
All suitability-related commanrs return a \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Targets related to the suitability of a site for affordable
housing and/or early learning facility development.
}
