% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_crtieria.R
\name{make_suitability_criteria}
\alias{make_suitability_criteria}
\alias{make_criteria_tax_exempt}
\alias{make_criteria_max_water_overlap_pct}
\alias{make_criteria_within_uga}
\alias{make_criteria_developable_zoning}
\alias{make_criteria_undevelopable_present_use}
\alias{make_criteria_lot_size}
\alias{make_criteria_area_ratio}
\alias{make_criteria_steep_vacant}
\alias{make_criteria_unbuildable}
\alias{make_criteria_other}
\title{Site Suiteability Criteria}
\usage{
make_suitability_criteria(criteria_tax_exempt, criteria_max_water_overlap_pct,
  criteria_within_uga, criteria_developable_zoning,
  criteria_undevelopable_presentuse, criteria_lot_size, criteria_area_ratio,
  criteria_steep_vacant, criteria_unbuildable, criteria_other)

make_criteria_tax_exempt()

make_criteria_max_water_overlap_pct()

make_criteria_within_uga()

make_criteria_developable_zoning(development_assumptions_zoning)

make_criteria_undevelopable_present_use()

make_criteria_lot_size(lot_size_breaks)

make_criteria_area_ratio()

make_criteria_steep_vacant()

make_criteria_unbuildable()

make_criteria_other()
}
\value{
All suitability criteria commands return a named list.
}
\description{
Targets related to the suitability \emph{criteria} of a site for affordable
housing and/or early learning facility development.
}
