% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit_stops_osm.R
\name{prepare_transit_stops_osm}
\alias{prepare_transit_stops_osm}
\alias{make_transit_stops_osm}
\alias{view_transit_stops_osm}
\title{Transit Stops (Open Street Map)}
\usage{
prepare_transit_stops_osm(path)

make_transit_stops_osm(path)

view_transit_stops_osm()
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Transit stops within King County, Washington.
}
\note{
\itemize{
\item{Data source: \code{\link[osmdata:opq]{osmdata::opq()}}}
}
}
\keyword{internal}
