% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uga.R
\name{prepare_uga}
\alias{prepare_uga}
\alias{make_uga}
\title{Urban Growth Area (King County)}
\usage{
prepare_uga(path)

make_uga(path, king_county)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Desc
}
\note{
Data source: Name of the source (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/urban_growth_SHP.zip})
}
