% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilization.R
\name{make_utilization}
\alias{make_utilization}
\alias{make_seattle_utilization_ratio}
\alias{make_utilization_present}
\alias{make_utilization_lot_size}
\alias{make_utilization_potential}
\title{Site Utilization}
\usage{
make_utilization(suitability, utilization_present, utilization_potential,
  seattle_utilization_ratio)

make_seattle_utilization_ratio(parcel_sf_ready, seattle_dev_cap)

make_utilization_present(parcel_ready, building)

make_utilization_lot_size(parcel_ready, utilization_criteria)

make_utilization_potential(suitability, development_assumptions_lot,
  utilization_lot_size)
}
\value{
All utilization-related commands return a \code{\link[tibble:tbl_df]{tbl_df}} object.
}
\description{
Targets related to the current utilization of a site
(e.g., what exists on the site, how much of the site is developed, how many residential units are there, etc.)
}
