% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizations_criteria.R
\name{make_utilization_criteria}
\alias{make_utilization_criteria}
\alias{make_utilization_criteria_lot_size}
\alias{make_utilization_criteria_ratio}
\alias{make_utilization_criteria_ratio_bins}
\title{Site Utilization Criteria}
\usage{
make_utilization_criteria(utilization_criteria_lot_size,
  utilization_criteria_ratio, utilization_criteria_ratio_bins)

make_utilization_criteria_lot_size(city_block_sqft, lot_types)

make_utilization_criteria_ratio()

make_utilization_criteria_ratio_bins()
}
\value{
All utilization criteria commands return a named list.
}
\description{
Targets related to the utilization \emph{criteria} of a site.
}
