% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbodies.R
\name{prepare_wa_major_waterbodies}
\alias{prepare_wa_major_waterbodies}
\alias{make_wa_major_waterbodies}
\alias{make_kc_waterbodies}
\title{National Hydrography Dataset Major Waterbodies (Washington)}
\usage{
prepare_wa_major_waterbodies(path)

make_wa_major_waterbodies(path)

make_kc_waterbodies(wa_major_waterbodies, king_county)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Desc
}
\note{
Data source: Washington Department of Ecology GIS Data (\url{https://fortress.wa.gov/ecy/gispublic/DataDownload/ECY_WAT_NHDWAMajor.zip})
}
