% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_workflow_plan}
\alias{get_workflow_plan}
\title{Get the Workflow Plan}
\usage{
get_workflow_plan()
}
\value{
a \code{\link[drake:drake_plan]{plan} tibble}
}
\description{
The drake \code{\link[drake:drake_plan]{plan}}s for creating
the project's workflow.
}
