% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\name{prepare_zoning}
\alias{prepare_zoning}
\alias{make_zoning}
\title{Zoning (King County)}
\usage{
prepare_zoning(path)

make_zoning(path)
}
\value{
Returns an \code{\link[sf:sf]{sf object}}.
}
\description{
Consolidated zoning categories for King County, Washington.
}
\note{
\itemize{
\item{Data source: King County GIS (\url{ftp://ftp.kingcounty.gov/gis-web/GISData/zoning_kc_consol_20_SHP.zip})}
\item{Metadata: \url{http://www5.kingcounty.gov/sdc/Metadata.aspx?Layer=zoning_kc_consol_20}}
}
}
