% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct.R
\name{make_acct}
\alias{make_acct}
\title{Real Property Account Extract}
\usage{
make_acct(path)
}
\description{
Account information for owners of parcels in King County.
}
\note{
Data source: King County Assessor's Office (restricted access)
}
