% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-restrictions.R
\name{make_env_restrictions}
\alias{make_env_restrictions}
\title{Parcel Environmental Restrictions}
\usage{
make_env_restrictions(path)
}
\description{
Environmental restrictions of parcels in King County.
}
\note{
Data source: King County Assessor Data Portal
}
